refactor: useConnectionsListQuery - move to hooks folder - prep for headless - decouple from InstallIntegrationContextProvider
